+++ ntex/src/server/service.rs
@@ -84,7 +84,7 @@ where
     fn call(&self, (guard, req): (Option<CounterGuard>, ServerMessage)) -> Self::Future {
         match req {
             ServerMessage::Connect(stream) => {
-                let stream = FromStream::from_stdstream(stream).map_err(|e| {
+                let stream = FromStream::from_stream(stream).map_err(|e| {
                     error!("Can not convert to an async tcp stream: {}", e);
                 });